asdasd

+++ weapon.cpp
@@ -50,7 +50,7 @@ void Weapon::render()
 	switch (_state)
 	{
 	case ST_ON_GROUND:
-		sVideo->renderGraphicEx(_spriteID, _Rect.x - sCamera->getX(), _Rect.y - sCamera->getY(), _Rect.w - 2, _Rect.h - 2, 1 + _state, 1 + _Rect.h * _weaponType + _weaponType, _Rect.w * 1.5f, _Rect.h * 1.5f, rotation, _Rect.w / 2, _Rect.w / 2, 0);
+		sVideo->renderGraphicEx(_spriteID, _Rect.x - sCamera->getX(), _Rect.y - sCamera->getY(), _Rect.w - 2, _Rect.h - 2, 1 + _state, 1 + _Rect.h * _weaponType + _weaponType, (int) _Rect.w * 1.5f, (int) _Rect.h * 1.5f, rotation, _Rect.w / 2, _Rect.w / 2, 0);
 		break;
 	default:
 		break;
@@ -69,7 +69,7 @@ void Weapon::renderInventory(int posX, int posY, double angulo, int rotated)
 	case ST_EQUIPED:
 		sVideo->renderGraphic(_spriteID, 0, 0, _Rect.w, _Rect.h, _Rect.w * 2 + 2, _Rect.h * 2 + 2); // Meter en el hud
 
-		sVideo->renderGraphicEx(_spriteID, posX, posY - _Rect.h, _Rect.w - 4, _Rect.h - 2, 1 + _state, 1 + _Rect.h * _weaponType + _weaponType, _Rect.w * 1.5f, _Rect.h * 1.5f, angulo + 180, 3, 20, rotated);
+		sVideo->renderGraphicEx(_spriteID, posX, posY - _Rect.h, _Rect.w - 4, _Rect.h - 2, 1 + _state, 1 + _Rect.h * _weaponType + _weaponType, (int) _Rect.w * 1.5f, (int) _Rect.h * 1.5f, angulo + 180, 3, 20, rotated);
 		break;
 	default:
 		break;
